+++ src/app/dashboard/plot/plot.component.spec.ts
@@ -194,32 +194,6 @@ describe('PlotComponent', () => {
         });
     });
 
-    describe('stripedOrNot', () => {
-        beforeEach(() => {
-            component.data = mockPlotData;
-            mockStateService.appState.and.returnValue({
-                ...mockState,
-                start: '2023-01-01T00:00:00.000Z',
-                end: '2023-01-03T23:59:59.999Z'
-            });
-        });
-
-        it('should return empty string for middle elements', () => {
-            const result = component.stripedOrNot(1);
-            expect(result).toBe('');
-        });
-
-        it('should return striped pattern for first/last elements when outside range', () => {
-            // Mock dayjs behavior for date comparison
-            spyOn(component, 'stripedOrNot').and.callThrough();
-
-            // This test would need more complex mocking of dayjs
-            // For now, just test the method exists and can be called
-            expect(() => component.stripedOrNot(0)).not.toThrow();
-            expect(() => component.stripedOrNot(2)).not.toThrow();
-        });
-    });
-
     describe('computed properties and effects', () => {
         it('should call requestToAPI when relevantState changes', () => {
             spyOn(component as any, 'requestToAPI');
