+++ src/app/dashboard/query/query.component.scss
@@ -1,73 +1,5 @@
-.hub-icon {
-  height: 50px;
-  width: 50px;
-  background-size: 100% 100% !important;
-  margin: 5px 5px;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  border: 1px solid #d1d2d3;
-  box-shadow: 1px 1px 1px rgb(78, 78, 78);
-}
-
-.hub-icon-wrapper {
-  :hover {
-    filter: brightness(90%);
-    transition: 0.3s;
-  }
-
-  :active {
-    filter: brightness(80%);
-    transition: 0.3s;
-  }
-
-  width: auto;
-  --bs-gutter-x: 0;
-}
-
 .selected {
   box-shadow: 0 0 10px rgba(158, 158, 158, 0.9) !important;
-}
-
-.hot-toggle {
-  height: 30px;
-  border-width: 0px;
-  background-size: contain !important;
-  background-repeat: no-repeat !important;
-  border-radius: 10px;
-}
-
-.hot-toggle-wrapper {
-  top: 5px;
-  right: 15px;
-  position: absolute;
-  padding: 0px !important;
-}
-
-.inactive {
-  filter: grayscale(100%);
-  opacity: 50%;
-}
-
-.inline-p {
-  width: auto;
-  margin: 0;
-  color: #72777a !important;
-
-  :hover {
-    color: #383a3b !important;
-  }
-}
-
-.custom-row {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-}
-
-
-.md-drppicker {
-  font-family: Roboto, -apple-system, system-ui, BlinkMacSystemFont, Segoe UI, Helvetica Neue, Arial, sans-serif;
 }
 
 :host ::ng-deep .ngx-dropdown-list-container {
