+++ src/app/dashboard/query/query.component.spec.ts
@@ -1,4 +1,4 @@
-import {ComponentFixture, fakeAsync, TestBed} from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 import {FormsModule} from '@angular/forms';
 import {QueryComponent} from './query.component';
 import {DataService} from '../../data.service';
@@ -100,8 +100,6 @@ describe('QueryComponent', () => {
     describe('Component Initialization', () => {
         it('should initialize with default values', () => {
             expect(component.interval).toBe(undefined);
-            expect(component.liveMode).toBe(false);
-            expect(component.hot_controls).toBe(false);
             expect(component.selectedCountries).toEqual([]);
             expect(component.selectedTopics).toEqual([]);
         });
@@ -193,19 +191,6 @@ describe('QueryComponent', () => {
             getElementByIdSpy.calls.reset();
         });
 
-
-        it('should fail validation when date range input is empty', () => {
-            const result = component.validateForm();
-
-            expect(result).toBe(false);
-            expect(mockToastService.show).toHaveBeenCalledWith({
-                title: 'Date range is empty',
-                body: 'Please provide a valid Date range',
-                type: 'error',
-                time: 5000
-            });
-        });
-
         it('should fail validation when selectedDateRangeUTC is undefined', () => {
             component.selectedDateRange = undefined;
 
@@ -243,7 +228,7 @@ describe('QueryComponent', () => {
         it('should not proceed if validation fails', () => {
             spyOn(component, 'validateForm').and.returnValue(false);
 
-            component.updateStateToFromSelection();
+            component.updateStateFromSelection();
 
             expect(mockStateService.updatePartialState).not.toHaveBeenCalled();
         });
@@ -252,7 +237,7 @@ describe('QueryComponent', () => {
             spyOn(component, 'validateForm').and.returnValue(true);
             spyOn(component, 'cleanHashTag').and.returnValue('missingmaps');
             component.interval = "P1M"
-            component.updateStateToFromSelection();
+            component.updateStateFromSelection();
 
             expect(mockStateService.updatePartialState).toHaveBeenCalledWith(jasmine.objectContaining({
                 hashtag: 'missingmaps',
@@ -266,7 +251,7 @@ describe('QueryComponent', () => {
 
             component.selectedCountries = component.dropdownOptions;
 
-            component.updateStateToFromSelection();
+            component.updateStateFromSelection();
 
             expect(component.countries).toEqual(['']);
         });
@@ -290,66 +275,6 @@ describe('QueryComponent', () => {
             const hashtagObj = {hashtag: '#test-hashtag', highlighted: ''};
             const result = component.cleanHashTag(hashtagObj);
             expect(decodeURIComponent(result)).toBe('test-hashtag');
-        });
-    });
-
-    describe('Hub Selection', () => {
-        beforeEach(() => {
-            component.dropdownOptions = [
-                {name: 'Afghanistan', value: 'AFG'},
-                {name: 'Bangladesh', value: 'BGD'},
-                {name: 'Brunei', value: 'BRN'},
-                {name: 'United States', value: 'USA'}
-            ];
-        });
-
-        it('should select countries for asia-pacific hub', () => {
-            component.changeHub('asia-pacific');
-
-            expect(component.selectedHub).toBe('asia-pacific');
-            expect(component.selectedCountries.length).toBeGreaterThan(0);
-            expect(component.selectedCountries.some(c => c.value === 'AFG')).toBe(true);
-            expect(component.selectedCountries.some(c => c.value === 'BGD')).toBe(true);
-        });
-
-        it('should select countries for other hubs', () => {
-            const hubs = ['la-carribean', 'wna', 'esa'];
-
-            hubs.forEach(hub => {
-                component.changeHub(hub);
-                expect(component.selectedHub).toBe(hub);
-                expect(Array.isArray(component.selectedCountries)).toBe(true);
-            });
-        });
-    });
-
-    describe('Impact Area Selection', () => {
-        beforeEach(() => {
-            component.topicOptions = [
-                {name: 'Water & Sanitation', value: 'wash'},
-                {name: 'Waterways', value: 'waterway'},
-                {name: 'Social Facilities', value: 'social_facility'},
-                {name: 'Places', value: 'place'},
-                {name: 'Land Use', value: 'lulc'}
-            ];
-        });
-
-        it('should select topics for disaster impact area', () => {
-            component.changeImpactArea('disaster');
-
-            expect(component.selectedImpactArea).toBe('disaster');
-            expect(component.selectedTopics.length).toBeGreaterThan(0);
-            expect(component.selectedTopics.some(t => t.value === 'wash')).toBe(true);
-        });
-
-        it('should select topics for all impact areas', () => {
-            const impactAreas = ['disaster', 'sus_cities', 'pub_health', 'migration', 'g_equality'];
-
-            impactAreas.forEach(area => {
-                component.changeImpactArea(area);
-                expect(component.selectedImpactArea).toBe(area);
-                expect(Array.isArray(component.selectedTopics)).toBe(true);
-            });
         });
     });
 
@@ -406,61 +331,6 @@ describe('QueryComponent', () => {
     });
 
 
-    describe('Live Mode', () => {
-        beforeEach(() => {
-            component.selectedDateRange = {
-                start: dayjs().subtract(2, 'hours'),
-                end: dayjs()
-            };
-            component.interval = 'PT5M';
-        });
-
-        it('should enable live mode button when conditions are met', () => {
-            const canEnable = component.enableLiveModeButton();
-            expect(canEnable).toBe(true);
-        });
-
-        it('should not enable live mode button when interval is not PT5M', () => {
-            component.interval = 'P1D';
-
-            const canEnable = component.enableLiveModeButton();
-            expect(canEnable).toBe(false);
-        });
-
-        it('should toggle live mode on', fakeAsync(() => {
-
-            spyOn(component, 'triggerMetaDataRetrieval');
-
-            component.toggleLiveMode();
-
-            expect(component.liveMode).toBe(true);
-            expect(component.triggerMetaDataRetrieval).toHaveBeenCalled();
-            expect(mockDataService.toggleLiveMode).toHaveBeenCalledWith(true);
-        }));
-
-        it('should toggle live mode off', () => {
-            component.liveMode = true;
-            spyOn(component, 'turnOffLiveMode');
-
-            component.toggleLiveMode();
-
-            expect(component.turnOffLiveMode).toHaveBeenCalled();
-        });
-
-        it('should turn off live mode completely', () => {
-            component.liveMode = true;
-            component.refreshIntervalId = 123;
-            spyOn(window, 'clearInterval');
-
-            component.turnOffLiveMode();
-
-            expect(component.liveMode).toBe(false);
-            expect(mockDataService.toggleLiveMode).toHaveBeenCalledWith(false);
-            expect(clearInterval).toHaveBeenCalledWith(123);
-        });
-    });
-
-
     describe('State Updates', () => {
         it('should update form from state correctly', () => {
             const inputData: IStateParams = {
@@ -490,18 +360,6 @@ describe('QueryComponent', () => {
             fixture.detectChanges();
 
             expect(component.interval).toBe('P1W');
-        });
-    });
-
-    describe('Component Cleanup', () => {
-        it('should cleanup on destroy', () => {
-            spyOn(component, 'turnOffLiveMode');
-            component['subscription'] = jasmine.createSpyObj('Subscription', ['unsubscribe']);
-
-            component.ngOnDestroy();
-
-            expect(component['subscription'].unsubscribe).toHaveBeenCalled();
-            expect(component.turnOffLiveMode).toHaveBeenCalled();
         });
     });
 });
