+++ src/app/data.service.ts
@@ -3,7 +3,6 @@ import {HttpClient} from '@angular/common/http';
 import {BehaviorSubject, catchError, map, Observable, retry, tap, throwError} from 'rxjs';
 
 import {environment} from '../environments/environment';
-import {ActivatedRoute} from '@angular/router';
 import {
     HexDataType,
     IHashtag,
@@ -43,7 +42,6 @@ export class DataService {
 
     constructor(
         private http: HttpClient,
-        private route: ActivatedRoute
     ) {
     }
 
@@ -129,5 +127,4 @@ export class DataService {
     toggleLiveMode(mode: boolean) {
         this.bsLive.next(mode)
     }
-
 }
